# Updated UI on both screen, also updated cell UI. # Refactoring, cleaning.

+++ trading-app/trading-app/data/network/Responses.swift
@@ -16,7 +16,7 @@ struct FinnResponse: Codable {
 struct ResponseData: Codable, Hashable {
     
     var ca: [String]
-    var price: Float
+    var price: Double
     var id: String
     var time: Double
     var value: Int
